pyproject.toml
[tool.poetry]
name = "league_piper"
version = "0.1.4"
description = "get league of legends related information and analysis in seconds"
authors = ["Ruoyu Zhu"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.2"
numpy = "^1.23.5"
requests = "^2.28.1"
pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
myst-nb = "^0.17.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
